v1.4.3: Fix critical regression + 15 more issues from 4th review round
CRITICAL (regression from v1.4.2):
- #5: messages_to_hy3 passed ChatMessage object to _content_to_str instead of
  .content — sent Pydantic repr as prompt on every tool-call round-trip and
  every request whose last message isn't from the user. Now correctly uses
  prior[i].content.

HIGH (security & robustness):
- #6: hmac.compare_digest(str, str) raises TypeError on non-ASCII tokens.
  Added _constant_time_match() that encodes to bytes. Also removed any()
  short-circuit so comparison count doesn't leak matched key position.
- #7: API_KEYS=' , ' parsed to empty set — silently ran open. Now warns at
  startup if API_KEYS was set but parsed to zero valid keys.
- #16: acquire() didn't catch CancelledError (client disconnect while queued)
  — record never finalized. Now re-raises CancelledError cleanly.
- #17: release() underflow guard silently leaked semaphore slots on
  double-release. Added _slot context manager for structural ownership.
- #18: Per-request httpx.AsyncClient → shared client via FastAPI lifespan.
  Eliminates ~100-300ms TCP+TLS handshake per request.
- #28: hy3.sh: 0 inside if ! body was always 0 (reflected if, not curl).
  Now captures curl_rc before the if.
- #29: hy3.sh: first curl had no error handling. Now wraps with exit code
  capture and Python JSON extraction with try/except.

MEDIUM:
- #19: HTTP_TIMEOUT now configurable via HTTP_READ_TIMEOUT env var (default 90).
- #20: Replaced manual SSE buffer += chunk (O(n²)) with aiter_lines().
  Also removes unbounded buffer risk.
- #21: record.has_history now set from actual history list, not message count.
- #22: Type annotations fixed: Optional[int] → Optional[float] for latency
  fields (upstream_post_latency_ms, upstream_stream_latency_ms, queued_ms).
- #23: PRESERVED_THINKING now env-tunable via _parse_bool_env.
- #30: hy3.sh: TOOLS_JSON file-vs-inline now validated — typo'd filename
  gives clear error instead of Python traceback.
- #31: hy3.sh: trap now covers EXIT INT TERM (was EXIT only).

LOW:
- #8: n validator documented — n=None accepted as 'not specified', n>1 → 422.
- #9: parse_hy3_data: extracted _safe_idx helper for clarity.
- #10: Added comment documenting Hy3 cumulative-snapshot invariant.
- #24: Removed dead system_prompt = '' initial assignment.
- #25: .dockerignore: removed wrong docs/ reference.
- #26: render.yaml: PORT/HOST now consistently quoted strings.
- #32: hy3.sh: removed unused tool_calls variable in heredoc.

Test battery: 23/23 passed (Phase 5). All 106 test executions across 5 phases
passed with zero regressions.

+++ logging_layer.py
@@ -126,8 +126,8 @@ class RequestRecord:
         # Upstream Hy3 details
         self.upstream_event_id: Optional[str] = None
         self.upstream_post_status: Optional[int] = None
-        self.upstream_post_latency_ms: Optional[int] = None
-        self.upstream_stream_latency_ms: Optional[int] = None
+        self.upstream_post_latency_ms: Optional[float] = None
+        self.upstream_stream_latency_ms: Optional[float] = None
         self.upstream_chunks: int = 0
         # Response details
         self.response_chars: int = 0
@@ -136,7 +136,7 @@ class RequestRecord:
         self.finish_reason: Optional[str] = None
         # Concurrency at the time of request
         self.concurrency_active_at_start: Optional[int] = None
-        self.queued_ms: Optional[int] = None
+        self.queued_ms: Optional[float] = None
 
     def to_dict(self) -> dict:
         return {
